Angler changes
If a furniture drops from vanilla mechanics, it will be converted into 1 of the 3 potions. Afterward, the mod rolls for furniture (without potions) with 6 / 19 chance of failing (in pre-1.4.4 it was a 57 out of 70 chance or returning a potion instead of furniture).

A bottomless bucket is guaranteed on the 25th quest completed (in addition to anything else that was rewarded).